Some debug logging in queue example. Hack to fix malloc in threads.
Malloc calls sbrk which uses the stack pointer to see if we're low on memory.
The problem is when we allocate the task stacks in .data they're below the heap
and so to sbrk we're totally out of memory. If I place the task stacks on the
main stack things work much better.