Adjust priority of SysTick_IRQn to fix scheduling issues when masked.
So this took some chasing... When using SEGGER RTT stuff the default locking
mechanism will ask IRQs above 0x20. What this means is that when sending a ton
of data over the RTT mechanism, esp if blocking for the FIFO to drain we end up
in long periods of masked IRQs.... so, with that range masked and SysTick at
0x00 we could fire multiple times before the PendSV gets a chance! This means
we'd get called w/o having actually changed tasks.