Disable IRQs in some status checks to chase down status issues.
My theory is that because some of these checks are done in task space,
the task space loop will find itself on a task during an iteration and
then preemption to a place where that task changes status. Then when we
return to the task we check the new status.

Gonna disable IRQs in those places and then eventually move the calls to
use service calls.