Fix infinite loop in find_new_task.
Right now my theory is  that the loop ends up on the "readings" task,
which is running (Idle) and because that task is lower priority than the
network task and of different priority we end up jumping to the head of
the queue, which happens to be the reading task. So, I've decided to
change the loop so that instead we continue the iteration the same way
we continue when we come upon a task of equal priority:

With this in place, I'm thinking the loop will be able to break on the
ondition at the end when we circle back around to the task after the
running task.

https://gitlab.com/fieldkit/firmware/-/issues/16